+++ app/src/main/java/com/unsa/suppliers/ui/view/auth/LoginFragment.kt
@@ -13,7 +13,7 @@ import com.google.android.material.textfield.TextInputEditText
 import com.google.android.material.textfield.TextInputLayout
 import com.unsa.suppliers.R
 import com.unsa.suppliers.databinding.FragmentLoginBinding
-import com.unsa.suppliers.ui.view.suppliers.SupplierActivity
+import com.unsa.suppliers.ui.view.main.MainActivity
 import com.unsa.suppliers.ui.viewmodel.auth.AuthViewModel
 
 class LoginFragment : Fragment() {
@@ -31,7 +31,7 @@ class LoginFragment : Fragment() {
         initListeners()
         authViewModel.token.observe(viewLifecycleOwner) {
             Toast.makeText(context, "JWT: ${authViewModel.token.value}", Toast.LENGTH_SHORT).show()
-            val intent = Intent(activity, SupplierActivity::class.java)
+            val intent = Intent(activity, MainActivity::class.java)
             startActivity(intent)
             (activity as AuthActivity).finish()
         }
