+++ app/src/main/java/com/unsa/suppliers/ui/view/main/MainActivity.kt
@@ -1,0 +1,23 @@
+package com.unsa.suppliers.ui.view.main
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.widget.Toast
+import androidx.activity.viewModels
+import com.unsa.suppliers.databinding.ActivityMainBinding
+import com.unsa.suppliers.ui.viewmodel.main.MainViewModel
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
+    private val mainViewModel: MainViewModel by viewModels()
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        mainViewModel.getToken()
+        Toast.makeText(this, mainViewModel.getToken(), Toast.LENGTH_SHORT).show()
+        binding.supplierJwt.text = mainViewModel.getToken()
+    }
+}
