+++ app/src/main/java/com/unsa/suppliers/ui/viewmodel/main/MainViewModel.kt
@@ -1,7 +1,6 @@
-package com.unsa.suppliers.ui.viewmodel.suppliers
+package com.unsa.suppliers.ui.viewmodel.main
 
 import android.util.Log
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.unsa.suppliers.data.PreferencesRepository
@@ -10,7 +9,7 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class SupplierViewModel @Inject constructor (
+class MainViewModel @Inject constructor (
     private val preferencesRepository: PreferencesRepository
 ) : ViewModel() {
     fun getToken(): String {
